+++ schemas/base-metadata.schema.ts
@@ -1,5 +1,7 @@
-import { serverGlobals, resolve } from '../src/renderer/src/server'
+import { serverGlobals, resolve } from '../src/renderer/src/server/globals'
+
 import { header } from '../src/renderer/src/stories/forms/utils'
+
 import baseMetadataSchema from './json/base_metadata_schema.json' assert { type: "json" }
 
 function getSpeciesNameComponents(arr: any[]) {
