+++ src/renderer/src/globals.js
@@ -1,4 +1,4 @@
-import { path, port } from "./electron/index.js";
+import { path } from "./electron/index.js";
 
 import guideGlobalMetadata from "../../../guideGlobalMetadata.json" assert { type: "json" };
 
@@ -8,8 +8,5 @@ export let runOnLoad = (fn) => {
     if (document.readyState === "complete") fn();
     else window.addEventListener("load", fn);
 };
-
-// Base Request URL for Python Server
-export const baseUrl = `http://127.0.0.1:${port}`;
 
 export const supportedInterfaces = guideGlobalMetadata.supported_interfaces;
