+++ src/renderer/src/stories/pages/uploads/UploadsPage.js
@@ -26,8 +26,8 @@ import { header } from "../../forms/utils";
 import { validateDANDIApiKey } from "../../../validation/dandi";
 import { InfoBox } from "../../InfoBox.js";
 
-import { onServerOpen } from "../../../server";
-import { baseUrl } from "../../../globals.js";
+import { onServerOpen } from "../../../server/index.js";
+import { baseUrl } from "../../../server/globals.js";
 
 export const isStaging = (id) => parseInt(id) >= 100000;
